pyproject.toml
[project]
name = "lol-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.3.3",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "selenium>=4.36.0",
]
